refactor(goal): 시뮬레이션을 GoalService로 통합하고 forecast 조립 중복 제거
월 저축액 시뮬레이션은 저축액만 바꿔 진단하는 것과 같아 별도 서비스를 두지 않고
진단이 있는 GoalService로 합쳤다. GoalSimulationService는 삭제.

예상 달성 시점 조립은 GoalForecastResponse.of() 정적 팩토리로 내려
상세 조회와 시뮬레이션이 같은 코드를 지나도록 했다.
