+++ src/main/java/com/team/independence/goal/service/GoalDetailServiceImpl.java
@@ -18,8 +18,9 @@ import org.springframework.stereotype.Service;
 
 /**
  * 목표 상세 화면 데이터를 모아서 반환한다.
- * 예상 달성 시점(forecasts)은 목표 진단과 같은 복리 계산을 쓴다 —
- * 월 저축액만 기준별로 바꿔 넣고 목표 금액에 닿는 시점을 GoalService에서 구한다.
+ * 예상 달성 시점(forecasts)은 저축액 세 개에 대한 시뮬레이션이다 —
+ * 계산은 GoalService가, 응답 조립은 GoalForecastResponse가 맡아
+ * 시뮬레이션 API와 같은 금액에 같은 날짜를 낸다.
  */
 @Service
 @RequiredArgsConstructor
@@ -31,7 +32,7 @@ public class GoalDetailServiceImpl implements GoalDetailService {
     private final GoalHousingMapper goalHousingMapper;
     private final SavingRecordMapper savingRecordMapper;
     private final AssetService assetService;
-    private final GoalService goalService; // 진단과 동일한 복리 계산 재사용
+    private final GoalService goalService; // 목표 조회·소유권 검증, 도달 개월수 계산
 
     @Override
     public GoalDetailResponse getGoalDetail(Long memberId, Long goalId) {
@@ -44,6 +45,7 @@ public class GoalDetailServiceImpl implements GoalDetailService {
 
         long targetAmount = goal.getTargetAmount();
         long remainingAmount = Math.max(0, targetAmount - currentAmount);
+
 
         List<SavingRecord> recentRecords = savingRecordMapper.findRecentByGoalId(goalId, RECENT_MONTHS);
         GoalDetailResponse.SavingStatus savingStatus = buildSavingStatus(goal, recentRecords);
@@ -61,7 +63,7 @@ public class GoalDetailServiceImpl implements GoalDetailService {
                         .achievementRate(calculateAchievementRate(currentAmount, targetAmount))
                         .build())
                 .savingStatus(savingStatus)
-                .forecasts(buildForecasts(netWorth, targetAmount, remainingAmount, savingStatus))
+                .forecasts(buildForecasts(netWorth, targetAmount, savingStatus))
                 .build();
     }
 
@@ -116,21 +118,20 @@ public class GoalDetailServiceImpl implements GoalDetailService {
     /**
      * 저축 기준별 예상 달성 시점.
      * 기준 금액이 없거나 0 이하면 계산이 불가하므로 목록에서 뺀다.
-     * monthsDiff는 고정 저축액 기준과 비교한 값이라, 고정 기준을 못 구하면 null이 된다.
+     * 항목 하나를 만드는 일은 시뮬레이션 API와 공유해야 해서 GoalService에 맡긴다.
      */
     private List<GoalForecastResponse> buildForecasts(AssetNetWorthBreakdown netWorth,
                                                       long targetAmount,
-                                                      long remainingAmount,
                                                       GoalDetailResponse.SavingStatus savingStatus) {
         Long fixedMonths = calculateMonths(netWorth, savingStatus.getFixedSaving(), targetAmount);
 
         List<GoalForecastResponse> forecasts = new ArrayList<>();
         addForecast(forecasts, SavingBasis.FIXED, savingStatus.getFixedSaving(),
-                netWorth, targetAmount, remainingAmount, fixedMonths);
+                netWorth, targetAmount, fixedMonths);
         addForecast(forecasts, SavingBasis.RECENT_AVERAGE, savingStatus.getRecentAverageSaving(),
-                netWorth, targetAmount, remainingAmount, fixedMonths);
+                netWorth, targetAmount, fixedMonths);
         addForecast(forecasts, SavingBasis.LATEST, savingStatus.getLatestSaving(),
-                netWorth, targetAmount, remainingAmount, fixedMonths);
+                netWorth, targetAmount, fixedMonths);
         return forecasts;
     }
 
@@ -139,30 +140,12 @@ public class GoalDetailServiceImpl implements GoalDetailService {
                              Long monthlySaving,
                              AssetNetWorthBreakdown netWorth,
                              long targetAmount,
-                             long remainingAmount,
                              Long fixedMonths) {
         if (monthlySaving == null || monthlySaving <= 0) {
             return;
         }
-
-        // 이미 달성했으면 예상 시점을 따질 게 없다.
-        if (remainingAmount == 0) {
-            forecasts.add(GoalForecastResponse.builder()
-                    .basis(basis)
-                    .monthlySaving(monthlySaving)
-                    .expectedDate(null)
-                    .monthsDiff(0)
-                    .build());
-            return;
-        }
-
-        Long months = calculateMonths(netWorth, monthlySaving, targetAmount);
-        forecasts.add(GoalForecastResponse.builder()
-                .basis(basis)
-                .monthlySaving(monthlySaving)
-                .expectedDate(months == null ? null : YearMonth.now().plusMonths(months))
-                .monthsDiff(fixedMonths == null || months == null ? null : (int) (fixedMonths - months))
-                .build());
+        Long months = goalService.calculateMonthToReach(netWorth, monthlySaving, targetAmount);
+        forecasts.add(GoalForecastResponse.of(basis, monthlySaving, months, fixedMonths));
     }
 
     private Long calculateMonths(AssetNetWorthBreakdown netWorth, Long monthlySaving, long targetAmount) {
