+++ src/main/java/com/team/independence/goal/service/GoalService.java
@@ -5,6 +5,7 @@ import com.team.independence.goal.domain.Goal;
 import com.team.independence.goal.dto.GoalCreateRequest;
 import com.team.independence.goal.dto.GoalDiagnosisRequest;
 import com.team.independence.goal.dto.GoalDiagnosisResponse;
+import com.team.independence.goal.dto.GoalForecastResponse;
 import com.team.independence.goal.dto.GoalResponse;
 
 public interface GoalService {
@@ -25,4 +26,10 @@ public interface GoalService {
      * 이미 도달했으면 0, 저축액이 0 이하거나 탐색 상한까지 못 미치면 null.
      */
     Long calculateMonthToReach(AssetNetWorthBreakdown netWorth, long monthlySaving, long targetAmount);
+
+    /**
+     * 월 저축액을 monthlySaving으로 바꿨다고 가정했을 때의 예상 달성 시점을 계산한다. 저장하지 않는다.
+     * 저축액이 0 이하면 GOAL_INVALID_INPUT, 진행 중이 아닌 목표면 GOAL_NOT_ACTIVE.
+     */
+    GoalForecastResponse simulateMonthlySaving(Long memberId, Long goalId, Long monthlySaving);
 }
